Update cellMath.py
fixed x variable, should be fully working at this point. Will function  with "wire" objects as inputs and in the heuristics file there is a method to combine the a[0]-a[3] vals of a gate and one input wire and assign it to the wire. That wire should be used as input for the math functions